refactor(#3038): run_bot S5 closing pass — final extraction + ratchet accounting
Measurements before S5 extraction (S4 HEAD): run_bot was 302 inclusive source lines, 370 production LoC in runtime_bootstrap.rs. Remaining inline blocks were utility-skip guard 20 lines, pre-shared bootstrap/settings/skills/voice 41, restored state + SharedData + relay 61, cached token + voice + intake worker 22, gateway lease acquire/skip 21, and one cohesive leader gateway runtime tail of 120 lines. That tail was the only remaining contiguous >=80-100-line phase block.

Extraction: moved the leader gateway runtime tail to runtime_bootstrap/gateway_runtime.rs. The moved body preserves startup order: restored generation/model/fast-mode logging -> health registry register -> slash command/framework/client build -> gateway lease keepalive spawn -> SIGTERM handler spawn -> gateway backend run. Token proof: compared against HEAD, the helper body is character-identical and token-identical after excluding only the wrapper signature (119 moved body lines, 275 non-whitespace tokens).

Post-S5 accounting: run_bot is 204 inclusive source lines; runtime_bootstrap.rs ratchets 369 -> 274 production LoC. gateway_runtime.rs lands at 147 raw/prod lines. runtime_bootstrap child cap check: max child is recovery_flush.rs at 357 raw/prod lines, all children remain under the 700-line namespace cap. Remaining inline root blocks are all below the extraction threshold; largest is restored state + SharedData + relay at 61 lines, so no extra extraction was forced.

Spawn distribution unchanged: runtime_bootstrap.rs/root run_bot has 0 direct tokio::spawn sites. Direct runtime_bootstrap child tokio::spawn sites total 12: spawns.rs 5, shutdown.rs 2, and gateway_lease.rs/intake.rs/orphan_recovery.rs/recovery_flush.rs/voice.rs 1 each.

Test discovery/gates: cargo test --lib runtime_bootstrap matched services::discord::runtime_bootstrap::{bootstrap_tests,restart_lifecycle_characterization_tests,session_gc::thread_session_gc_tests} (13 tests). cargo test --lib run_bot matched 0 tests. cargo test --lib discord umbrella passed with 1577 passed / 0 failed. Gates passed before commit: cargo fmt --check; cargo check --lib; cargo clippy --lib -- -D warnings; cargo test --lib runtime_bootstrap; cargo test --lib run_bot; cargo test --lib discord 2>/dev/null | tail -1; python3 scripts/audit_maintainability.py --check.
