+++ src/services/discord/runtime_bootstrap.rs
@@ -2,6 +2,7 @@ use super::*;
 
 mod framework_setup;
 mod gateway_lease;
+mod gateway_runtime;
 mod intake;
 mod orphan_recovery;
 mod queued_placeholders;
@@ -18,6 +19,7 @@ use self::framework_setup::{run_bot_build_slash_commands, run_bot_framework_setu
 use self::gateway_lease::{
     GatewayLeaseOutcome, run_bot_acquire_gateway_lease, run_bot_spawn_gateway_lease_keepalive,
 };
+use self::gateway_runtime::run_bot_start_gateway_runtime;
 use self::intake::run_bot_maybe_spawn_intake_worker;
 #[allow(unused_imports)]
 pub(in crate::services::discord) use self::queued_placeholders::{
@@ -240,123 +242,25 @@ pub(crate) async fn run_bot(token: &str, provider: ProviderKind, context: RunBot
         }
     };
 
-    {
-        let ts = chrono::Local::now().format("%H:%M:%S");
-        tracing::info!(
-            "  [{ts}] 🔑 dcserver generation: {}",
-            shared.restart.current_generation
-        );
-        if !restored_model_overrides.is_empty() {
-            tracing::info!(
-                "  [{ts}] 🧩 restored model overrides: {} channel(s)",
-                restored_model_overrides.len()
-            );
-        }
-        if !restored_fast_mode_channels.is_empty() {
-            tracing::info!(
-                "  [{ts}] ⚡ restored fast mode channels: {} channel(s)",
-                restored_fast_mode_channels.len()
-            );
-        }
-    }
-
-    // Register this provider with the health check registry
-    health_registry
-        .register(provider.as_str().to_string(), shared.clone())
-        .await;
-
-    let token_owned = token.to_string();
-    let shared_clone = shared.clone();
-    let voice_config_for_setup = voice_config.clone();
-    let voice_receiver_for_setup = voice_receiver.clone();
-
-    let slash_commands = run_bot_build_slash_commands();
-
-    let framework = poise::Framework::builder()
-        .options(poise::FrameworkOptions {
-            commands: slash_commands,
-            command_check: Some(|ctx| {
-                Box::pin(async move {
-                    let settings_snapshot = { ctx.data().shared.settings.read().await.clone() };
-                    let allowed = provider_handles_channel(
-                        ctx.serenity_context(),
-                        &ctx.data().provider,
-                        &settings_snapshot,
-                        ctx.channel_id(),
-                    )
-                    .await;
-                    if !allowed {
-                        let ts = chrono::Local::now().format("%H:%M:%S");
-                        tracing::info!(
-                            "  [{ts}] ⏭ CMD-GUARD: skipping /{} in channel {} for provider {}",
-                            ctx.command().name,
-                            ctx.channel_id(),
-                            ctx.data().provider.as_str()
-                        );
-                    }
-                    Ok(allowed)
-                })
-            }),
-            event_handler: |ctx, event, _framework, data| Box::pin(handle_event(ctx, event, data)),
-            ..Default::default()
-        })
-        .setup(move |ctx, _ready, framework| {
-            let shared_for_migrate = shared_clone.clone();
-            let health_registry_for_setup = health_registry.clone();
-            let provider_for_setup = provider_for_framework.clone();
-            let token_for_ready = token_owned.clone();
-            let voice_config_for_setup = voice_config_for_setup.clone();
-            let voice_receiver_for_setup = voice_receiver_for_setup.clone();
-            Box::pin(run_bot_framework_setup(
-                ctx,
-                _ready,
-                framework,
-                shared_for_migrate,
-                shared_clone,
-                health_registry_for_setup,
-                provider_for_setup,
-                token_for_ready,
-                token_owned,
-                voice_config_for_setup,
-                voice_receiver_for_setup,
-                startup_reconcile_remaining,
-                startup_doctor_started,
-                api_port,
-            ))
-        })
-        .build();
-
-    let intents = discord_gateway_intents();
-
-    let client = commands::register_songbird(serenity::ClientBuilder::new(token, intents))
-        .framework(framework)
-        .await
-        .expect("Failed to create Discord client");
-
-    let gateway_lease_task = gateway_lease.map(|lease| {
-        run_bot_spawn_gateway_lease_keepalive(
-            lease,
-            &shared,
-            &provider,
-            client.shard_manager.clone(),
-        )
-    });
-
-    // Graceful shutdown: on SIGTERM, persist queue/inflight/last_message state
-    // and quick-exit. tmux/TUI processes survive — the next dcserver instance
-    // rehydrates the channel bindings (see rehydrate_existing_claude_tui_bindings;
-    // polled every CLAUDE_IDLE_REHYDRATE_POLL_INTERVAL ≈ 5s) and resumes transcript
-    // tailing from the persisted last_offset.
-    run_bot_spawn_sigterm_handler(&shared, provider_for_shutdown);
-
-    run_bot_run_gateway_backend(
-        client,
-        &provider_for_error,
-        gateway_lease_task,
+    run_bot_start_gateway_runtime(
+        token,
+        provider,
+        provider_for_error,
+        provider_for_framework,
+        provider_for_shutdown,
+        startup_reconcile_remaining,
+        startup_doctor_started,
+        health_registry,
         startup_reconcile_remaining_for_client_start,
         startup_doctor_started_for_client_start,
         health_registry_for_client_start,
         api_port,
+        shared,
+        voice_config,
+        voice_receiver,
+        gateway_lease,
+        &restored_model_overrides,
+        &restored_fast_mode_channels,
     )
     .await;
 }
